Fix for regression test failure
The underlying failure was due to a spatial pooler change. The result of that is similar to changing the random seed so it's basically random variation. Here I updated minThreshold to be more in line with other models and make this model a bit more robust. This brings it back in line with previous performance.

Fix line wrap issue.

Fixes #35